refactor(icons): Rename ApprovalStatusClass
Rename ApprovalStatusClass to ApprovalStatusIcons and move the file to shared directory.
Split Pending filter into two separate filters for ORG_APPROVED and PENDING.
Basic Code formating fixes and bugfixes.

Closes AMLNG-215 AMLNG-820
