+++ app/js/stores/DiscoveryPageStore.js
@@ -99,7 +99,7 @@ var DiscoveryPageStore = Reflux.createStore({
         arr = this.sortAlphabetically(arr);
         arr.sort(function (a, b) {
             if (a.avgRate == b.avgRate) {
-                return (a.totalVotes < b.totalVotes) ? -1 : 1;
+                return (a.totalReviews < b.totalReviews) ? -1 : 1;
             } else {
                 return (a.avgRate < b.avgRate) ? -1 : 1;
             }
