+++ app/styles/management/user/_my-listings.scss
@@ -113,8 +113,8 @@
         }
     }
 
-    .label-draft, .label-pending, .label-needs-action, .label-published, .label-rejected, .label-pending-delete {
-        i { 
+    .label-draft, .label-pending, .label-needs-action, .label-published, .label-rejected, .label-pending-delete, .label-pending-center, .label-pending-org, .label-pending-deletion {
+        i {
             float: left;
         }
