+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "center-ui",
-  "version": "1.1.94",
+  "version": "1.1.95",
   "repository": {
     "type": "git",
     "url": "https://github.com/ozone-development/ozp-center.git"
